+++ test/oscope/test_runner.clj
@@ -20,6 +20,7 @@
             [oscope.sample-emitter-test]
             [oscope.server-test]
             [oscope.telemetry-test]
+            [oscope.typed-schema-test]
             [oscope.visualization-editor-test]
             [oscope.web-test]
             [oscope.workbench-test]))
@@ -38,6 +39,7 @@
                         'oscope.query-view-test 'oscope.raw-export-test
                         'oscope.sample-emitter-test
                         'oscope.server-test 'oscope.visualization-editor-test
-                        'oscope.telemetry-test 'oscope.web-test
+                        'oscope.telemetry-test 'oscope.typed-schema-test
+                        'oscope.web-test
                         'oscope.workbench-test)]
     (System/exit (if (zero? (+ fail error)) 0 1))))